requirements.txt
ghsearcher
tweepy
opensearch-py
google-cloud-translate
iso-639
pyyaml
jsonpath-ng